realtime-translation/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.1
distro==1.9.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
llvmlite==0.45.0
MarkupSafe==3.0.2
multidict==6.6.4
numba==0.62.0
numpy==2.3.3
openai==1.107.3
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
resampy==0.4.3
sniffio==1.3.1
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
